requirements.txt
requests==2.7.0
watson_developer_cloud==1.3.3
Django==2.0.5
python_twitter==3.4.1
